Restore previous behaviour on burn failure.
The previous behaviour on failure during signing the burn transaction
was to delete the chomp results that had been created. These hanging
results seem to result in issues when coming back to reveal. I'm
restoring the previous behaviour temporarily but want to return to this
issue once I've accounted for all code paths: I think ideally we should
not be deleting results.